feat: replace app with webview
